requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Type checking
mypy>=1.0.0
pandas-stubs>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
